Expose functions for editable & scripts packages
The editablePackage and scriptsPackage elements of the mkPoetryEnv
function are wholly separable and really ought to be moved into the
functionality of the mkPoetryPackages function. By doing so both will
automatically be available in the mkPoetryApplication's dependencyEnv as
well.

I thought the functionality that encapsulates these might further see
potential use even outside the context of mkPoetryPackages, so I made
them top-level functions and documented them in the README.