pep508: Fix `in` operator
If a marker contains two or more instances of the substring "in" Nix
regex engine in it's infitine wisdom splits only on the last one.

By changing a * to a + we can get the desired behaviour, but we do
need to fall back to the previous behaviour in case the more
restrictive match is not working.

Closes #332