overrides: Move build-systems to separate file which can be automatically generated
This change uses the new development tool to derive build-systems from
nixpkgs to the extent that it is possible.

This will reduce the maintenance burden on us a bit until the upstream Poetry issue around this is solved.
Additionally it will fix a lot of packages we didn't even know were broken to begin with.