fix: avoid infinite recursion while searching for path
In `findGitIgnores`, to find the parent path, we do `path + "/.."`. This means that, for each iteration, the path name grows:

1. `/nix/store/1234-example/a/b/`
2. `/nix/store/1234-example/a/b/..`
3. `/nix/store/1234-example/a/b/../..`
4. etc.

Thus, checking if path is not equal to `/` will always return `true`.

With this fix, it will instead check that the path exists. If the path goes higher than `/`, it will not exist, thus fixing the false positive.

@moduon MT-83