feat: Show lock location in package errors
Tested by temporarily setting `meta.broken = true;` in
`mk-poetry-dep.nix`, and then

    $ nix-build tools -A env
    error: Package ‘python3.9-click-8.1.3’ in /home/user/h/poetry2nix/tools/poetry.lock:0 is marked as broken, refusing to evaluate.

       a) To temporarily allo[...]

As you can see, the location of the lock file is reported, which
is especially useful in Nixpkgs, which contains multiple lockfiles.

This used to be just the location of a generic python builder, which
is generally not what the user wants.

I've exposed the `poetrylockPos` variable as a parameter that callers
may set. This is intended as a contingency more so than a feature that
is useful to users. That's why I've decided not to document it,
because it would just add noise.