issue template: remove bullet list
A bullet list would make sense
if the code blocks would be part of the items.
To make a code block part of a list item in Markdown,
it would though have to be indented
which is either too laborious (when indenting all lines)
or too confusing (when indenting only the opening line).
If they aren't indented, they'd interrupt the list
(thus effectively splitting it into 3 one-item lists)
so let's instead just have the file names
as prose-ish captions of the code blocks.