feat: autofix version for all packages that use setuptools-scm
Building a package that has setuptools-scm usually fails if it doesn't define an override like the one I'm removing in `pyyaml-include`.

Or sometimes it works, but packages that depend on it don't work because they define some lower bound of the dependency, but the dependency doesn't have an appropriate version defined.

With this simple fix, all packages that use setuptools-scm will include the `SETUPTOOLS_SCM_PRETEND_VERSION` to let them inherit the appropriate version.

@moduon MT-1075