mkPoetryEnv: add checkGroups
Currently it is not possible to remove the "dev" group when using mkPoetryEnv: the "dev" group is still built because it is part of the `checkGroups`. To be able to avoid this, I opted to add `checkGroups` to `mkPoetryEnv`, just like what you would be able to in `mkPoetryApplication` and `mkPoetryPackages`.