Refer to right pkg-config thing
In #1447 some references of `pkgs.pkgconfig` were
change to `pkgs.pkg-config` which was correct, as
the `pkgs.pkgconfig` is a deprecated alias for the
latter. However that change _also_ changed
`self.pkgconfig` to `self.pkg-config`: the `self`
here is _not_ the same as `pkgs`. The `self`
refers to Python package set so it formerly
referred to the `pkgconfig` Python package.
There's no `pkg-config` Python package so this
code just fails to evaluate, resulting in #1564.

In first place, it looks like using
`self.pkgconfig` was probably wrong and we wanted
the `pkgs.pkg-config` so I simply make that change
instead.

Fixes #1564.