pyside6: ensure PySide6 can be imported and used
The PySide6/__init__.py script is at runtime trying to find the
dependent Qt libraries in subdirectories of its path in the installed
site-packages directory. This fails because the dependent
pyside6-essentials and pyside6-addons packages aren't installed into the
same site-packages directory. As a work-around this commit symlinks
those files into the site-packages directory of the `pyside6` package.

This commit also adds a small application that imports
and verifies that PySide6 can be used after building.